folder-gateway-skill/ai-task-worker/requirements.txt
celery
redis
pyyaml
sqlalchemy
